+++ routes/Artists.js
@@ -6,22 +6,11 @@ exports.list = function (req, res) {
         //    res.error(e)
         //})
 };
-//exports.listSongs = function (req, res) {
-//    var name = req.param("name");
-//    global.db.selectAllSongsByArtistName(name)
-//        .then(function (data) {
-//            res.send(data);
-//        }, function (e) {
-//            res.error(e)
-//        })
-//};
-
-
 
 exports.listPaginated = function (req, res) {
 
-    var index = parseInt(req.param("index"));
-    var length = parseInt(req.param("length"));
+    var index = parseInt(req.params.index);
+    var length = parseInt(req.params.length);
 
     var response = {
         length: global.artists.length,