+++ ui/routes/SongQueue/SongQueue.js
@@ -4,10 +4,8 @@ define(['jquery', 'handlebars', 'text!./template.html'], function ($, Handlebars
         this.element = $.parseHTML('<div></div>')[0];
         var template = Handlebars.compile(this.html);
 
-        this.songTemplte = Handlebars.compile(this.songHtml);
-        //console.log(options);
-
         this.element.innerHTML = template(options);
+        $(this.element).i18n();
     }
 
     SongQueue.prototype.addSong = function (song) {
@@ -22,26 +20,20 @@ define(['jquery', 'handlebars', 'text!./template.html'], function ($, Handlebars
         $('#queue-song-current tbody').children().remove();
 
 
-        $(createTr(song)).hide().appendTo(this.element.querySelector('#queue-song-current tbody')).fadeIn(1000);
+        $(createTr.call(this, song)).hide().appendTo(this.element.querySelector('#queue-song-current tbody')).fadeIn(1000);
     };
 
     function createTr(song) {
         var tr = document.createElement("TR");
         tr.setAttribute("data-id", song.id);
         var title = tr.insertCell(0);
-        title.innerHTML = (song.title) ? song.title : 'Unknown';
+        title.innerHTML = (song.title) ? song.title : this.options.strings.unknown;
         var album = tr.insertCell(1);
-        album.innerHTML = (song.album) ? song.album : 'Unknown';
+        album.innerHTML = (song.album) ? song.album : this.options.strings.unknown;
         var artist = tr.insertCell(2);
-        artist.innerHTML = (song.artist) ? song.artist : 'Unknown';
-
+        artist.innerHTML = (song.artist) ? song.artist : this.options.strings.unknown;
         return tr;
     }
-
-    SongQueue.prototype.songHtml =
-        '<td>{{#if title}}{{title}}{{else}}(Unknown){{/if}}</td>' +
-        '<td>{{#if album}}{{album}}{{else}}(Unknown){{/if}}</td>' +
-        '<td>{{#if artist}}{{artist}}{{else}}(Unknown){{/if}}</td>';
 
     SongQueue.prototype.html = html;